// App-FilimuFy/src/app/interfaces/serie/tv.interface.ts
export interface TVShowListResponse {
    page: number
    results: TVShow[]
    total_pages: number
    total_results: number
  }
  
  export interface TVShow {
    adult: boolean
    backdrop_path?: string
    genre_ids: number[]
    id: number
    origin_country: string[]
    original_language: string
    original_name: string
    overview: string
    popularity: number
    poster_path?: string
    first_air_date: string
    name: string
    vote_average: number
    vote_count: number
  }
  